fixes #30
