+++ app/controllers/v1/micropost_participants_controller.rb
@@ -1,0 +1,15 @@
+module V1
+  class MicropostParticipantsController < ApplicationController
+    def trollers
+      @trollers = Troller.valid_ones
+
+      respond_with @trollers
+    end
+
+    def targets
+      @targets = Target.valid_ones
+
+      respond_with @targets
+    end
+  end
+end
